Split build and install steps.
Few people flagged it to me already. I think it make sense to build
release before installing with only caviat of having to `sudo make
clean` later on. Since I don't have a strong opinion about it, I'll
switch back to the conventional way of doing it, that is to split build
and install steps.

Closes #10